% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_data.R
\docType{data}
\name{earthquake_data}
\alias{earthquake_data}
\title{earthquake data}
\format{
## `earthquake_data`
Datasets contain records of 782 earthquakes from 1/1/2001 to 1/1/2023. The meaning of all columns is as follows::
\describe{
  \item{title}{title name given to the earthquake}
  \item{magnitude}{The magnitude of the earthquake}
  \item{date_time}{date and time}
  \item{cdi}{The maximum reported intensity for the event range}
  \item{mmi}{The maximum estimated instrumental intensity for the event}
  \item{alert}{The alert level - “green”, “yellow”, “orange”, and “red”}
  \item{tsunami}{"1" for events in oceanic regions and "0" otherwise}
  \item{sig}{A number describing how significant the event is. Larger numbers indicate a more significant event. This value is determined on a number of factors, including: magnitude, maximum MMI, felt reports, and estimated impact}
  \item{net}{The ID of a data contributor. Identifies the network considered to be the preferred source of information for this event.}
  \item{nst}{The total number of seismic stations used to determine earthquake location.}
  \item{dmin}{Horizontal distance from the epicenter to the nearest station}
  \item{gap}{The largest azimuthal gap between azimuthally adjacent stations (in degrees). In general, the smaller this number, the more reliable is the calculated horizontal position of the earthquake. Earthquake locations in which the azimuthal gap exceeds 180 degrees typically have large location and depth uncertainties}
  \item{magType}{The method or algorithm used to calculate the preferred magnitude for the event}
  \item{depth}{The depth where the earthquake begins to rupture}
  \item{latitude / longitude}{coordinate system by means of which the position or location of any place on Earth's surface can be determined and described}
  \item{location}{location within the country}
  \item{continent}{continent of the earthquake hit country}
  \item{country}{affected country}
}
}
\usage{
earthquake_data
}
\description{
this is from kaggle public dataset, which is used for more analysis with map together
}
\references{
<https://www.kaggle.com/datasets/warcoder/earthquake-dataset/data>
}
\keyword{datasets}
