% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishy_data.R
\docType{data}
\name{fishy_data}
\alias{fishy_data}
\title{Example Fish observation data}
\format{
300 rows, 1 column
\describe{\item{x}{This is the column that has the fish name of the observed fish (character)}}
}
\source{
Naomi Tague \url{gauchospace.ucsb.edu}
}
\usage{
fishy_data
}
\description{
These are fish observations from Naomi Tague. This can be summarized using the \code{fish_summary} function in this package.
}
\keyword{datasets}
