% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamenmap.R
\name{generateStamenMap}
\alias{generateStamenMap}
\title{stmenmap is to build a stamenmap.}
\usage{
generateStamenMap(address, maptype, zoom)
}
\arguments{
\item{address}{An address, e.g. New York City}

\item{maptype}{The map type that stamen supported.}

\item{zoom}{The zoom parameter is an integer between 0 (zoomed out) and 18 (zoomed in). 18 is normally the maximum.}
}
\value{
An image of stamen map type.
}
\description{
generateStamenMap is to build a stamen type map with a given address, map type and zoom range.
}
\examples{
generateStamenMap("ryd linkoping", "watercolor", 15)
}
\references{
Reference page link <https://maps.stamen.com/#watercolor/11/40.6575/-73.9685>
}
