% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_fxn.R
\name{growth_fxn}
\alias{growth_fxn}
\title{growth_fxn - a simple population growth equation}
\usage{
growth_fxn(a, r, t, m, P = 0)
}
\arguments{
\item{a}{Size of initial population}

\item{r}{Growth rate of population of interest}

\item{t}{Time in years}

\item{m}{Maximum population}
}
\value{
P The final population size at time t + n, organized in a 2x2
data frame
}
\description{
This simple exponential growth formula can be used to calculate population size at a given time.
}
\details{
Use this formula to calculate a final population size, using any give initial population, growth rate, and time.
}
