% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest_while_fun.R
\name{interest_fun}
\alias{interest_fun}
\title{interest_fun - a while loop function for calculating compound interest}
\source{
compound interest formula:
 https://www.thecalculatorsite.com/articles/finance/compound-interest-formula.php#:~:text=Compound%20interest%2C%20or%20'interest%20on,the%20number%20of%20time%20periods
}
\usage{
interest_fun(balance, target, interest_rate, compound_per_yr, yr)
}
\arguments{
\item{balance}{the amount of money in a bank account. When setting this
parameter, you are setting the amount of initial balance. The function then
produces the final balance after applying compound interest for X years}

\item{interest_rate}{the rate at which interest is being applied}

\item{compound_per_yr}{the amount of times the balance is compounded per year.
This typically should be set to 1 most of the time, but can be changed by
user in situations where that is not the case}

\item{yr}{number of years since the initial investment was made, the amount of
time that balance has had time to gain interest. Starts with year 0 as the
first year}
}
\description{
Purpose: to calculate the final balance in a bank account and amount of years
it would take to reach a certain target in an account where an initial
investment earns compound interest. The function uses the compound interest
formula and contains a `while()` loop which will continue to apply interest on
a given investment until a given target balance is reached. It will report the
amount of years it would take to reach that target balance and the actual
final balance that is achieved at the end of the year in which the target was
reached.
}
