% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_chi_square.R
\name{print.chisq_result}
\alias{print.chisq_result}
\title{Print method for chisq_result objects}
\usage{
\method{print}{chisq_result}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"chisq_result"}.}

\item{...}{Further arguments passed to or from other methods (ignored).}
}
\value{
\code{x}, invisibly.
}
\description{
Displays a formatted summary of the chi-square test of independence results.
}
