% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_ttest.R
\name{print.ttest_result}
\alias{print.ttest_result}
\title{Print method for ttest_result objects}
\usage{
\method{print}{ttest_result}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"ttest_result"}.}

\item{...}{Further arguments passed to or from other methods (ignored).}
}
\value{
\code{x}, invisibly.
}
\description{
Displays a formatted summary of the two-sample t-test results.
}
