% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ricker.R
\name{ricker}
\alias{ricker}
\title{ricker - a discrete population model}
\usage{
ricker(A0, J0, u, y0, a, b, c11, c12, c21, c22)
}
\arguments{
\item{A0}{Adult count of species x at timestep 0}

\item{J0}{Juvenile count of species x at timestep 0 J}

\item{u}{Juvenile death rate must be between 0 and 1}

\item{y0}{Total population of species x at timestep 0}

\item{a}{Reproduction coefficient for species x}

\item{b}{Reproduction coefficient for species y}

\item{c11}{Self competition coefficient for species x}

\item{c12}{Interspecies competition coefficient for species x}

\item{c21}{Self competition coefficient for species y}

\item{c22}{Interspecies competition coefficient for species y}
}
\value{
timeseries A dataframe with the timestep, the population of species x and species y.
}
\description{
This is an implementation of the adult/juvinile ricker
population model.
}
\details{
This function models the population dynamics of two competing
species, one that has a life phase difference and the other that doesn't.
This function uses a Adult/Juvinile Ricker model, a discrete
competition model and therefore can benefit from the use of a while loop.
\subsection{Modules:}{
\itemize{
\item Initialize the outputed dataframe
\item Use initial counts for first timestep calculation
\item Loop discrete function iteration
\item Store and return results at the end
}
}
}
\references{
Cushing, J.M., Sheree Levarge, Nakul Chitnis, and Shandelle M.
Henson. "Some Discrete Competition Models and the Competitive Exclusion
Principle." Journal of Difference Equations and Applications 10, no.
13–15 (November 2004): 1139–51.
https://doi.org/10.1080/10236190410001652739.
}
