\name{stamenmap-package}
\alias{stamenmap-package}
\alias{stamenmap}
\docType{package}
\title{
\packageTitle{stamenmap}
}
\description{
\packageDescription{stamenmap}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{stamenmap}
\packageIndices{stamenmap}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{stamenmap}

Maintainer: \packageMaintainer{stamenmap}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
