% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_add.R
\name{my_add}
\alias{my_add}
\title{Sum function}
\usage{
my_add(x, y = 10)
}
\arguments{
\item{x}{first number}

\item{y}{second number}
}
\description{
Sum the two arguments. The y parameter will have default value of 10. If one of
the two argument is NA, the functions return NA
}
