% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_add.R
\name{my_add}
\alias{my_add}
\title{Add two values}
\usage{
my_add(x, y = 10)
}
\arguments{
\item{x}{First interger}

\item{y}{Second interger}
}
\value{
A interger sum of x and y
}
\description{
Add two values
}
\examples{
my_add(10,20)
}
