% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_add.R
\name{my_add}
\alias{my_add}
\title{Add two numbers}
\usage{
my_add(x, y = 10)
}
\arguments{
\item{x}{A number.}

\item{y}{Another number.}
}
\value{
A number that is the sum of x and y
}
\description{
Add two numbers
}
\examples{
x <- 10
y <- 15
my_add(x, y)
}
