% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_add.R
\name{my_add}
\alias{my_add}
\title{My add}
\usage{
my_add(x, y = 10)
}
\arguments{
\item{x}{an integer or null}

\item{y}{an integer or null}
}
\value{
integer or string
}
\description{
My add
}
\examples{
my_add(1, 2)
my_add(1)
my_add(NA)
my_add(10, NA)
my_add("10", "20")
}
