% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamenmap-package.R
\docType{package}
\name{stamenmap-package}
\alias{stamenmap}
\alias{stamenmap-package}
\title{stamenmap: This package will create a stamen map}
\description{
This package will create a stamen map when given an address, stamen map type and zoom range.
}
\author{
\strong{Maintainer}: Xuan Developer \email{xuawa284@student.liu.se}

Other contributors:
\itemize{
  \item Ashen Developer \email{thawe276@student.liu.se} [contributor]
}

}
\keyword{internal}
